requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-google-genai==2.0.10
langchain-community==0.3.27
langgraph==0.5.2
google-generativeai==0.8.5
python-dotenv==1.0.1
python-multipart==0.0.20
pandas==2.2.2
rouge-score==0.1.2
tiktoken==0.7.0
